requirements.txt
Django>=2.2.8,<3.0
djangorestframework>=3.9.4,<3.10.0
mysqlclient>=1.4.6,<1.5.0
Pillow>=6.2.1,<6.3.0
